Top level layout for 'main' branch
Switching the yaml-spec repo's primary branch from `master` to `main`.

The old `master` content is under the `1.2` directory.

There are now separate directories for the various parts of defining
YAML.

There is a new top level ReadMe.md describing the layout.

Most of the files in this comit are empty placeholders.
They will be filled in (or removed) over time.